+++ verification-server-service/src/main/java/es/gob/radarcovid/verification/controller/GenerationController.java
@@ -11,13 +11,12 @@ package es.gob.radarcovid.verification.controller;
 
 import com.sun.istack.NotNull;
 import es.gob.radarcovid.common.annotation.Loggable;
-import es.gob.radarcovid.common.handler.RadarCovidExceptionHandler;
+import es.gob.radarcovid.common.handler.RadarCovidHandler;
 import es.gob.radarcovid.verification.api.CodesResultDto;
 import es.gob.radarcovid.verification.api.MessageResponseDto;
-import es.gob.radarcovid.verification.business.VerificationService;
+import es.gob.radarcovid.verification.business.GenerationService;
 import es.gob.radarcovid.verification.etc.Constants;
-import es.gob.radarcovid.verification.etc.OpenApiConstants;
-import es.gob.radarcovid.verification.security.JwtAuthorizationFilter;
+import es.gob.radarcovid.verification.security.CCAAAuthorizationUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -35,7 +34,6 @@ import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -43,7 +41,6 @@ import javax.validation.constraints.Min;
 import java.util.concurrent.Callable;
 
 @RestController
-@RequestMapping(GenerationController.GENERATE_ROUTE)
 @RequiredArgsConstructor
 @Validated
 @Slf4j
@@ -51,20 +48,17 @@ public class GenerationController {
 
     public static final String GENERATE_ROUTE = "/generate";
 
-    public static final String API_KEY_AUTH = "apiKeyAuth";
-
-    private final VerificationService service;
+    private final GenerationService service;
 
     @Value("${application.entities.codes.max}")
     private int maxCodes;
 
     @Loggable
-    @Secured("CCAA")
+    @Secured({Constants.AUTH_RADARCOVID, Constants.AUTH_GENERATION})
     @Operation(
             summary = "Provides n verification codes",
             description = "Generates n verification codes to be used by Autonomous Communities so they can provide codes to affected people (n <= 1000)",
-            security = @SecurityRequirement(name = API_KEY_AUTH),
-            hidden = OpenApiConstants.GENERATION_GET_CODES_OPERATION_HIDDEN
+            security = @SecurityRequirement(name = Constants.API_KEY_AUTH)
     )
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Codes generated", content = @Content(schema = @Schema(implementation = CodesResultDto.class))),
@@ -72,39 +66,25 @@ public class GenerationController {
             @ApiResponse(responseCode = "500", description = "Exception", content = @Content(schema = @Schema(implementation = MessageResponseDto.class)))
     })
     @GetMapping(
+            path = GENERATE_ROUTE,
             produces = MediaType.APPLICATION_JSON_VALUE
     )
     public Callable<ResponseEntity<?>> getCodes(Authentication authentication, @RequestParam("n") @NotNull @Min(1) Integer number) {
-        String ccaa = getCCAAFromAuthentication(authentication);
+        String ccaa = CCAAAuthorizationUtil.getCCAAFromAuthentication(authentication);
         MDC.put(Constants.TRACKING, "GET_CODES|CCAA:" + ccaa + ",NUMBER:" + String.valueOf(number));
         if (number != null && 0 < number && number <= maxCodes) {
-            CodesResultDto result = service.getCodes(isRadarCovidAuthentication(authentication), ccaa, number);
+            CodesResultDto result = service.getCodes(CCAAAuthorizationUtil.isRadarCovidAuthentication(authentication), ccaa, number);
             if (result != null && result.getCodes() != null && !result.getCodes().isEmpty()) {
                 log.info("Generados {} códigos para {}", number, ccaa);
-                return () -> {
-                    return ResponseEntity.ok(result);
-                };
+                return () -> ResponseEntity.ok(result);
             } else {
-                return () -> {
-                    return RadarCovidExceptionHandler.buildResponseMessage(HttpStatus.BAD_REQUEST, "");
-                };
+                return () -> RadarCovidHandler.buildResponseMessage(HttpStatus.BAD_REQUEST, "");
             }
         } else {
             String message = "Number of codes (" + number + ") must be between 1 and the maximum (" + maxCodes + ")";
             log.warn(message);
-            return () -> {
-                return RadarCovidExceptionHandler.buildResponseMessage(HttpStatus.BAD_REQUEST, message);
-            };
+            return () -> RadarCovidHandler.buildResponseMessage(HttpStatus.BAD_REQUEST, message);
         }
-
     }
 
-    private String getCCAAFromAuthentication(Authentication authentication) {
-        return isRadarCovidAuthentication(authentication) ?
-                authentication.getName().substring(6, 8) : authentication.getName().substring(5, 7);
-    }
-
-    private boolean isRadarCovidAuthentication(Authentication authentication) {
-        return authentication.getName().startsWith(JwtAuthorizationFilter.RADAR_PREFIX);
-    }
 }