+++ verification-server-service/src/main/java/es/gob/radarcovid/verification/persistence/impl/VerificationDaoImpl.java
@@ -10,9 +10,12 @@
 package es.gob.radarcovid.verification.persistence.impl;
 
 import es.gob.radarcovid.verification.persistence.VerificationDao;
+import es.gob.radarcovid.verification.persistence.entity.TanEntity;
 import es.gob.radarcovid.verification.persistence.entity.VerificationEntity;
+import es.gob.radarcovid.verification.persistence.repository.TanRepository;
 import es.gob.radarcovid.verification.persistence.repository.VerificationRepository;
 import es.gob.radarcovid.verification.util.HashingService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -23,59 +26,54 @@ import java.util.Date;
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 @Slf4j
 public class VerificationDaoImpl implements VerificationDao {
 
     private final VerificationRepository repository;
+    private final TanRepository tanRepository;
     private final HashingService hashingService;
 
-    public VerificationDaoImpl(
-            VerificationRepository repository,
-            HashingService hashingService) {
-        this.repository = repository;
-        this.hashingService = hashingService;
-    }
-
     @Override
-    public boolean isValidCode(String code) {
-        return repository.findByCodeHashAndCodeRedeemedIsFalse(hashingService.hash(code))
-                .filter(t -> t.codeCanBeRedeemed(LocalDateTime.now(ZoneOffset.UTC))).isPresent();
-    }
-
-    @Override
-    public boolean redeemCode(String code, String tan, Date validUntil) {
-        String tanHash = hashingService.hash(tan);
-        log.debug("Entering in redeemCode(code = {}, tan = {}, validUntil = {}): tanHash = {}", code, tan, validUntil, tanHash);
-        boolean result = repository.findByCodeHashAndCodeRedeemedIsFalse(hashingService.hash(code))
-                .filter(t -> t.codeCanBeRedeemed(LocalDateTime.now(ZoneOffset.UTC)))
-                .map(t -> {
-                    Date now = new Date();
-                    t.setCodeRedeemedAt(now);
-                    t.setTanHash(tanHash);
-                    t.setTanValidFrom(now);
-                    t.setTanValidUntil(validUntil);
-                    repository.save(t);
-                    return true;
-                }).map(t -> true).orElseGet(() -> false);
-        log.debug("Leaving redeemCode() with: {}", result);
-        return result;
-    }
+	public boolean redeemCode(String code, String tan, Date validUntil) {
+		String tanHash = hashingService.hash(tan);
+		String codeHash = hashingService.hash(code);
+		log.debug("Entering in redeemCode(code = {}, tan = {}, validUntil = {}): tanHash = {}", code, tan, validUntil,
+				tanHash);
+		
+		boolean result = repository.findByCodeHashAndCodeRedeemedIsFalse(codeHash)
+			.filter(v -> v.codeCanBeRedeemed(LocalDateTime.now(ZoneOffset.UTC)))
+			.map(v -> {
+				Date now = new Date();
+				v.setCodeRedeemedAt(now);
+				TanEntity tanEntity = TanEntity.builder().tanHash(tanHash).tanValidFrom(now)
+						.tanValidUntil(validUntil).codeHash(codeHash).build();
+				repository.save(v);
+				tanRepository.save(tanEntity);
+				return true;
+			}).orElseGet(() -> false);
+		log.debug("Leaving redeemCode() with: {}", result);
+		return result;
+	}
 
     @Override
     public boolean redeemTan(String tan) {
         log.debug("Entering in redeemTan(tan = {})", tan);
+        String tanHash = hashingService.hash(tan);
         LocalDateTime localDateTimeNow = LocalDateTime.now(ZoneOffset.UTC);
-        boolean result = repository.findByTanHashAndTanRedeemedIsFalse(hashingService.hash(tan))
-                .filter(t -> t.codeCanBeRedeemed(localDateTimeNow) && t.tanCanBeRedeemed(localDateTimeNow))
-                .map(t -> {
-                    Date now = new Date();
-                    t.setCodeRedeemed(true);
-                    t.setCodeRedeemedAt(now);
-                    t.setTanRedeemed(true);
-                    t.setTanRedeemedAt(now);
-                    repository.save(t);
-                    return true;
-                }).map(t -> true).orElseGet(() -> false);
+        
+        boolean result = tanRepository.findByTanHash(tanHash)
+			.filter(t -> t.tanCanBeRedeemed(localDateTimeNow))
+			.flatMap(t -> repository.findByCodeHashAndCodeRedeemedIsFalse(t.getCodeHash()))
+			.filter(v -> v.codeCanBeRedeemed(localDateTimeNow))
+            .map(v -> {
+            	Date now = new Date();
+                v.setCodeRedeemed(true);
+                v.setCodeRedeemedAt(now);
+                repository.save(v);
+                tanRepository.deleteByTanHash(tanHash);
+                return true;
+            }).orElseGet(() -> false);
         log.debug("Leaving redeemTan() with: {}", result);
         return result;
     }
@@ -97,4 +95,5 @@ public class VerificationDaoImpl implements VerificationDao {
         repository.saveAll(entities);
         return true;
     }
+    
 }